+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2.4.0
+
+### Added
+
+- New `Nanook::Node#confirmation_history` method.
+- New `Nanook::Block#confirm` method.
+- New `Nanook::Block#confirmed_recently?` method.
+
+### Changed
+
+- `Nanook::Block#generate_work` now can take optional `use_peers` argument.
+
 ## 2.3.0
 
 ### Added